+++ src/main/java/chess/service/GameService.java
@@ -49,11 +49,11 @@ public class GameService {
                 .orElseThrow(() -> new RuntimeException("찾는 게임이 존재하지 않습니다."));
     }
 
-    public List<GameResultDto> findHistoriesByMemberId(final Long memberId) {
-        final List<ChessGame> games = gameDao.findHistoriesByMemberId(memberId);
-        return games.stream()
-                .map(game -> toGameResultDTO(game, memberId))
-                .collect(Collectors.toList());
+    public List<ChessGame> findHistoriesByMemberId(final Long memberId) {
+        return gameDao.findHistoriesByMemberId(memberId);
+//        return games.stream()
+//                .map(game -> toGameResultDTO(game, memberId))
+//                .collect(Collectors.toList());
     }
 
     public void move(final Long gameId, final String rawFrom, final String rawTo) {
@@ -75,55 +75,6 @@ public class GameService {
         if (!game.getRoom().getPassword().equals(password)) {
             throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
         }
-    }
-
-    private GameResultDto toGameResultDTO(final ChessGame game, final Long memberId) {
-        final String winResult = findWinResult(game, memberId);
-        final String enemyName = findEnemyName(game, memberId);
-        final String team = findTeam(game, memberId);
-        final double myScore = findMyScore(game, memberId);
-        final double enemyScore = findEnemyScore(game, memberId);
-        return new GameResultDto(winResult, enemyName, team, myScore, enemyScore);
-    }
-
-    private static String findWinResult(final ChessGame game, final Long memberId) {
-        final Long winnerId = game.getWinnerId();
-        if (winnerId.equals(memberId)) {
-            return "승";
-        }
-        return "패";
-    }
-
-    private static String findEnemyName(final ChessGame game, final Long memberId) {
-        if (game.getParticipant().getBlackId().equals(memberId)) {
-            return game.getParticipant().getWhiteName();
-        }
-        return game.getParticipant().getBlackName();
-    }
-
-    private static String findTeam(final ChessGame game, final Long memberId) {
-        if (game.getParticipant().getBlackId().equals(memberId)) {
-            return "흑";
-        }
-        return "백";
-    }
-
-    private static double findMyScore(final ChessGame game, final Long memberId) {
-        final Result result = game.createResult();
-
-        if (game.getParticipant().getBlackId().equals(memberId)) {
-            return result.getBlackScore();
-        }
-        return result.getWhiteScore();
-    }
-
-    private static double findEnemyScore(final ChessGame game, final Long memberId) {
-        final Result result = game.createResult();
-
-        if (game.getParticipant().getBlackId().equals(memberId)) {
-            return result.getWhiteScore();
-        }
-        return result.getBlackScore();
     }
 
     private void updateGameByMove(final ChessGame chessGame, final String rawFrom, final String rawTo) {
