+++ src/test/java/chess/dao/MockGameDaoTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 
 class MockGameDaoTest {
 
-    private static final MockGameDao dao = new MockGameDao();
+    private final MockGameDao dao = new MockGameDao();
 
     @Test
     @DisplayName("체스 게임을 저장소에 저장한다.")
